First take on the actual app.
For now, the logic is in the UI app instead of the headless app. That
makes it easier during development to test if changes happens while the
app is down and so forth.

Right now, upon start of the app, it checks for existence and correct
permissions for the todo.txt file. If it exists, it copies the file into
the app data folder for caching purposes, unless it already existed
(does nothing in that case).

It then goes through all existing and valid notebooks that starts with
an '@'. Then creates a TodoList of the todo.txt file and creates new
notebooks as new unknown contexts are found in the TodoList.

That's it for now. But everything works smoothly so it's a great first
step.